+++ src/store.js
@@ -53,7 +53,7 @@
 WatcherMorph, Point, CustomBlockDefinition, Context, ReporterBlockMorph, Sound,
 CommandBlockMorph, detect, CustomCommandBlockMorph, CustomReporterBlockMorph,
 Color, List, newCanvas, Costume, Audio, IDE_Morph, ScriptsMorph, ArgLabelMorph,
-BlockMorph, ArgMorph, InputSlotMorph, TemplateSlotMorph, CommandSlotMorph,
+BlockMorph, ArgMorph, InputSlotMorph, TemplateSlotMorph, CommandSlotMorph, ZOOM,
 FunctionSlotMorph, MultiArgMorph, ColorSlotMorph, nop, CommentMorph, isNil,
 localize, SVG_Costume, MorphicPreferences, Process, isSnapObject, Variable,
 SyntaxElementMorph, BooleanSlotMorph, normalizeCanvas, contains, Scene,
@@ -63,7 +63,7 @@ Project, CustomHatBlockMorph, SnapVersion*/
 
 // Global stuff ////////////////////////////////////////////////////////
 
-modules.store = '2025-November-06';
+modules.store = '2025-December-20';
 
 // XML_Serializer ///////////////////////////////////////////////////////
 /*
@@ -345,6 +345,7 @@ SnapSerializer.prototype.loadProjectModel = function (
         app = appInfo ? appInfo.split(' ')[0] : null,
         appVersion = appInfo ? parseFloat(appInfo.split(' ')[1]) || 0 : 0,
         scenesModel = xmlNode.childNamed('scenes'),
+        zoom = xmlNode.attributes.zoom,
         project = new Project();
 
     if (ide && app && app !== this.app.split(' ')[0]) {
@@ -370,6 +371,9 @@ SnapSerializer.prototype.loadProjectModel = function (
         project.scenes.add(
             this.loadScene(xmlNode, appVersion, remixID, keepRoles)
         );
+    }
+    if (ide && zoom) {
+        ide.setZoom(+zoom, true); // no save
     }
     return project.initialize();
 };
@@ -2058,7 +2062,8 @@ Array.prototype.toXML = function (serializer) {
 
 Project.prototype.toXML = function (serializer) {
     var thumbdata,
-        scenes = this.scenes.asArray();
+        scenes = this.scenes.asArray(),
+        hasTemplate = scenes.some(any => any.role === 'template');
 
     // thumb data catch cross-origin tainting exception when using SVG costumes
     try {
@@ -2067,14 +2072,12 @@ Project.prototype.toXML = function (serializer) {
         thumbdata = null;
     }
 
-    if (scenes.some(any => any.createdFromTemplate) &&
-        !(scenes.some(any => any.role === 'template'))
-    ) {
+    if (scenes.some(any => any.createdFromTemplate) && !hasTemplate) {
         scenes = scenes.filter(each => each.role !== 'tutorial');
     }
 
     return serializer.format(
-        '<project name="@" app="@" version="@">' +
+        '<project name="@" app="@" version="@"%>' +
             '<notes>$</notes>' +
             '<thumbnail>$</thumbnail>' +
             '<scenes select="@">%</scenes>' +
@@ -2082,6 +2085,8 @@ Project.prototype.toXML = function (serializer) {
         this.name || localize('Untitled'),
         serializer.app,
         serializer.version,
+        hasTemplate && (ZOOM > 1) ?
+            ' zoom="' + Math.round(ZOOM * 100) + '"' : '',
         this.notes || '',
         thumbdata,
         scenes.indexOf(this.currentScene) + 1,